added LanguageItem component
