+++ src/components/Footer/index.jsx
@@ -1,19 +1,14 @@
 import style from './Footer.module.css';
 import { Link } from 'react-router-dom';
-import { useTranslation } from 'react-i18next';
+
 import logoImg from '../../assets/icons/logo.png';
 import callImg from '../../assets/icons/call.png';
 import messageImg from '../../assets/icons/message.png';
 import VButton from '../VButton';
 import { footerLinkRulesData, footerLinkPersonsData, footerSocialData } from './mockData';
+import LanguageItem from '../LanguageItem';
 
 export default function Footer() {
-    const { i18n } = useTranslation();
-
-    const handleLanguageChange = (lang) => {
-        i18n.changeLanguage(lang);
-    };
-
     return (
         <footer className={style.footer}>
             <div className="container">
@@ -25,15 +20,7 @@ export default function Footer() {
                             </Link>
                         </div>
                         <div className={style.languages}>
-                            <span className={i18n.language === 'UA' ? style.active : ''} onClick={() => handleLanguageChange('UA')}>
-                                UA
-                            </span>
-                            <span className={i18n.language === 'EN' ? style.active : ''} onClick={() => handleLanguageChange('EN')}>
-                                EN
-                            </span>
-                            <span className={i18n.language === 'DE' ? style.active : ''} onClick={() => handleLanguageChange('DE')}>
-                                DE
-                            </span>
+                            <LanguageItem />
                         </div>
                     </div>
                     <div className={style.center}>
