+++ src/components/LanguageItem/LanguageItem.module.css
@@ -1,0 +1,3 @@
+.active {
+  color: var(--font-color-thirdy);
+}
