+++ src/components/RulesForUsing/RulesForUsing.module.css
@@ -7,6 +7,7 @@
     margin: 0 auto;
     max-width: 780px;
     padding-bottom: 100px;
+    text-align: justify;
 }
 
 .header {
